Refactor out qos.rs
