+++ streamer/src/nonblocking/mod.rs
@@ -1,4 +1,5 @@
 pub mod connection_rate_limiter;
+pub mod qos;
 pub mod quic;
 #[cfg(feature = "dev-context-only-utils")]
 pub mod recvmmsg;
